requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
djangorestframework==3.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.6

